Seapring the algorithms into its own file
