+++ proyecto01/src/searchAlgorithms/breadthFirstSearch.go
@@ -1,0 +1,114 @@
+package searchAlgorithms
+
+import (
+	"fmt"
+	"math"
+	"time"
+
+	"github.com/Krud3/InteligenciaArtificial/src/datatypes"
+)
+
+type BreadthFirstSearch struct{}
+
+func (a *BreadthFirstSearch) LookForGoal(e *enviroment) datatypes.SearchResult {
+	var parentNodes []datatypes.AgentStep
+	// var paths [][]datatypes.AgentStep // To keep track of both the path to the passenger and from the passenger to the goal
+	expandenNodes := 0
+
+	initialPosition := e.agent.position
+	queue := datatypes.Queue[datatypes.AgentStep]{}
+	queue.Enqueue(initialPosition)
+
+	start := time.Now()
+
+	pathToPassenger := []datatypes.BoardCoordinate{} // Path from initial position to the passenger
+	pathToGoal := []datatypes.BoardCoordinate{}      // Path from passenger to the goal
+
+	for !queue.IsEmpty() {
+		currentStep, empty := queue.Dequeue()
+		if empty {
+			return datatypes.SearchResult{
+				PathFound:     []datatypes.BoardCoordinate{},
+				SolutionFound: false,
+				ExpandenNodes: expandenNodes,
+				TreeDepth:     0,
+				Cost:          0,
+				TimeExe:       time.Since(start),
+			}
+		}
+
+		parentNodes = append(parentNodes, currentStep)
+
+		// Phase 1: Find the passenger
+		if !e.agent.passenger && e.board[currentStep.CurrentPosition.X][currentStep.CurrentPosition.Y] == 5 {
+			e.passengerPosition = datatypes.BoardCoordinate{
+				X: currentStep.CurrentPosition.X,
+				Y: currentStep.CurrentPosition.Y,
+			}
+			e.agent.passenger = true // Mark the agent as having picked up the passenger
+
+			// Reconstruct path from the initial position to the passenger
+			pathToPassenger = reconstructPath(parentNodes, currentStep)
+
+			// Clear the queue and start BFS again from the passenger's position
+			queue.Clear()
+			parentNodes = nil // Clear parent nodes for the next phase
+			queue.Enqueue(
+				datatypes.AgentStep{
+					PreviousPosition: datatypes.BoardCoordinate{
+						X: math.MaxInt,
+						Y: math.MaxInt,
+					},
+					CurrentPosition: currentStep.CurrentPosition,
+					Depth:           currentStep.Depth + 1,
+					Action:          math.MaxInt,
+				},
+			) // Start from the passenger's position
+			continue
+		}
+
+		// Phase 2: Search for the goal (once the passenger is found)
+		if e.agent.passenger && e.board[currentStep.CurrentPosition.X][currentStep.CurrentPosition.Y] == 6 {
+			end := time.Now()
+			// Reconstruct path from the passenger to the goal
+			pathToGoal = reconstructPath(parentNodes, currentStep)
+
+			// Combine the two paths: initial -> passenger + passenger -> goal
+			combinedPath := append(pathToPassenger[1:], pathToGoal[1:]...)
+			var cost float32
+			for step := range combinedPath {
+				fmt.Println(getCellCost(e.board[combinedPath[step].X][combinedPath[step].Y]))
+				cost += float32(getCellCost(e.board[combinedPath[step].X][combinedPath[step].Y]))
+			}
+
+			return datatypes.SearchResult{
+				PathFound:     combinedPath,
+				SolutionFound: true,
+				ExpandenNodes: expandenNodes,
+				TreeDepth:     len(combinedPath),
+				Cost:          float32(cost), // Example cost, you can modify it as needed
+				TimeExe:       end.Sub(start),
+			}
+		}
+
+		// Continue exploring neighbors
+		e.agent.position = currentStep
+		agentPerception := Percept(e.agent, e.board)
+		expandenNodes++
+		for _, perception := range agentPerception {
+			cost := currentStep.Cost + int(getCellCost(e.board[perception.Coordinate.X][perception.Coordinate.Y]))
+			if (perception.Coordinate.X != currentStep.PreviousPosition.X) || (perception.Coordinate.Y != currentStep.PreviousPosition.Y) {
+				queue.Enqueue(
+					datatypes.AgentStep{
+						Action:           perception.Direction,
+						Depth:            currentStep.Depth + 1,
+						CurrentPosition:  perception.Coordinate,
+						PreviousPosition: currentStep.CurrentPosition,
+						Cost:             int(cost),
+					},
+				)
+			}
+		}
+	}
+	return datatypes.SearchResult{}
+}
