+++ proyecto01/src/searchAlgorithms/searchAlgorithms.go
@@ -176,115 +176,6 @@ func Percept(a agent, board [][]int) []datatypes.CoordinateMovement {
 	return canMove
 }
 
-type BreadthFirstSearch struct{}
-
-type UniformCostSearch struct{}
-
-type DepthSearch struct{}
-
-func (a *BreadthFirstSearch) LookForGoal(e *enviroment) datatypes.SearchResult {
-	var parentNodes []datatypes.AgentStep
-	// var paths [][]datatypes.AgentStep // To keep track of both the path to the passenger and from the passenger to the goal
-	expandenNodes := 0
-
-	initialPosition := e.agent.position
-	queue := datatypes.Queue[datatypes.AgentStep]{}
-	queue.Enqueue(initialPosition)
-
-	start := time.Now()
-
-	pathToPassenger := []datatypes.BoardCoordinate{} // Path from initial position to the passenger
-	pathToGoal := []datatypes.BoardCoordinate{}      // Path from passenger to the goal
-
-	for !queue.IsEmpty() {
-		currentStep, empty := queue.Dequeue()
-		if empty {
-			return datatypes.SearchResult{
-				PathFound:     []datatypes.BoardCoordinate{},
-				SolutionFound: false,
-				ExpandenNodes: expandenNodes,
-				TreeDepth:     0,
-				Cost:          0,
-				TimeExe:       time.Since(start),
-			}
-		}
-
-		parentNodes = append(parentNodes, currentStep)
-
-		// Phase 1: Find the passenger
-		if !e.agent.passenger && e.board[currentStep.CurrentPosition.X][currentStep.CurrentPosition.Y] == 5 {
-			e.passengerPosition = datatypes.BoardCoordinate{
-				X: currentStep.CurrentPosition.X,
-				Y: currentStep.CurrentPosition.Y,
-			}
-			e.agent.passenger = true // Mark the agent as having picked up the passenger
-
-			// Reconstruct path from the initial position to the passenger
-			pathToPassenger = reconstructPath(parentNodes, currentStep)
-
-			// Clear the queue and start BFS again from the passenger's position
-			queue.Clear()
-			parentNodes = nil // Clear parent nodes for the next phase
-			queue.Enqueue(
-				datatypes.AgentStep{
-					PreviousPosition: datatypes.BoardCoordinate{
-						X: math.MaxInt,
-						Y: math.MaxInt,
-					},
-					CurrentPosition: currentStep.CurrentPosition,
-					Depth:           currentStep.Depth + 1,
-					Action:          math.MaxInt,
-				},
-			) // Start from the passenger's position
-			continue
-		}
-
-		// Phase 2: Search for the goal (once the passenger is found)
-		if e.agent.passenger && e.board[currentStep.CurrentPosition.X][currentStep.CurrentPosition.Y] == 6 {
-			end := time.Now()
-			// Reconstruct path from the passenger to the goal
-			pathToGoal = reconstructPath(parentNodes, currentStep)
-
-			// Combine the two paths: initial -> passenger + passenger -> goal
-			combinedPath := append(pathToPassenger[1:], pathToGoal[1:]...)
-			var cost float32
-			for step := range combinedPath {
-				fmt.Println(getCellCost(e.board[combinedPath[step].X][combinedPath[step].Y]))
-				cost += float32(getCellCost(e.board[combinedPath[step].X][combinedPath[step].Y]))
-			}
-
-			return datatypes.SearchResult{
-				PathFound:     combinedPath,
-				SolutionFound: true,
-				ExpandenNodes: expandenNodes,
-				TreeDepth:     len(combinedPath),
-				Cost:          float32(cost), // Example cost, you can modify it as needed
-				TimeExe:       end.Sub(start),
-			}
-		}
-
-		// Continue exploring neighbors
-		e.agent.position = currentStep
-		agentPerception := Percept(e.agent, e.board)
-		expandenNodes++
-		for _, perception := range agentPerception {
-			cost := currentStep.Cost + int(getCellCost(e.board[perception.Coordinate.X][perception.Coordinate.Y]))
-			if (perception.Coordinate.X != currentStep.PreviousPosition.X) || (perception.Coordinate.Y != currentStep.PreviousPosition.Y) {
-				queue.Enqueue(
-					datatypes.AgentStep{
-						Action:           perception.Direction,
-						Depth:            currentStep.Depth + 1,
-						CurrentPosition:  perception.Coordinate,
-						PreviousPosition: currentStep.CurrentPosition,
-						Cost:             int(cost),
-					},
-				)
-			}
-		}
-	}
-	return datatypes.SearchResult{}
-}
-
 // Helper function to reconstruct the path from a given node back to the start
 func reconstructPath(parentNodes []datatypes.AgentStep, endStep datatypes.AgentStep) []datatypes.BoardCoordinate {
 	path := []datatypes.BoardCoordinate{}
@@ -315,116 +206,6 @@ func reconstructPath(parentNodes []datatypes.AgentStep, endStep datatypes.AgentS
 		}
 	}
 	return path
-}
-
-func (a *UniformCostSearch) LookForGoal(e *enviroment) datatypes.SearchResult {
-	// Step 1: Find the path to the passenger
-	start := time.Now()
-	pathToPassenger, passengerExpandedNodes, passengerCost := a.findPath(e, 5)
-	// If the path to the passenger is not found, return empty result
-	if len(pathToPassenger) == 0 {
-		return datatypes.SearchResult{
-			PathFound:     []datatypes.BoardCoordinate{},
-			SolutionFound: false,
-			ExpandenNodes: passengerExpandedNodes,
-			TreeDepth:     0,
-			Cost:          float32(passengerCost),
-			TimeExe:       time.Since(start),
-		}
-	}
-	// Step 2: Find the path from the passenger to the goal
-	pathToGoal, goalExpandedNodes, goalCost := a.findPath(e, 6)
-
-	// If the path to the goal is not found, return the path to the passenger
-	if len(pathToGoal) == 0 {
-		return datatypes.SearchResult{
-			PathFound:     pathToPassenger,
-			SolutionFound: false,
-			ExpandenNodes: goalExpandedNodes, // You might want to count these nodes as well
-			TreeDepth:     len(pathToPassenger),
-			Cost:          float32(goalCost),
-			TimeExe:       time.Since(start),
-		}
-	}
-
-	// Step 3: Combine the two paths
-	combinedPath := append(pathToPassenger, pathToGoal[1:]...) // Avoid duplicate passenger position
-
-	return datatypes.SearchResult{
-		PathFound:     combinedPath,
-		SolutionFound: true,
-		ExpandenNodes: goalExpandedNodes + passengerExpandedNodes, // You might want to count these nodes as well
-		TreeDepth:     len(combinedPath),
-		Cost:          float32(goalCost + passengerCost),
-		TimeExe:       time.Since(start),
-	}
-}
-
-// Helper function to perform Uniform Cost Search from start to goal
-func (a *UniformCostSearch) findPath(e *enviroment, goal int) ([]datatypes.BoardCoordinate, int, float32) {
-	var parentNodes []datatypes.AgentStep // Holds parent nodes
-	expandenNodes := 0
-	priorityQueue := datatypes.PriorityQueue[datatypes.AgentStep]{}
-
-	// Start with the initial position and cost 0
-	initialCost := float32(0)
-	priorityQueue.Push(datatypes.Element[datatypes.AgentStep]{
-		Value: datatypes.AgentStep{
-			Depth:  e.agent.position.Depth,
-			Action: e.agent.position.Action,
-			Cost:   int(initialCost),
-			PreviousPosition: datatypes.BoardCoordinate{
-				X: math.MaxInt,
-				Y: math.MaxInt,
-			},
-			CurrentPosition: e.agent.position.CurrentPosition,
-		},
-		Priority: 0}, // Priority can be based on cost from start
-	)
-
-	visited := make(map[datatypes.BoardCoordinate]bool)
-
-	for !priorityQueue.IsEmpty() {
-		currentStep, _ := priorityQueue.Pop()
-
-		// Check if the current position is the goal
-		if e.board[currentStep.CurrentPosition.X][currentStep.CurrentPosition.Y] == goal {
-			e.agent.position = currentStep
-			return reconstructPath(parentNodes, currentStep), expandenNodes, float32(currentStep.Cost)
-		}
-
-		if visited[currentStep.CurrentPosition] {
-			continue
-		}
-
-		visited[currentStep.CurrentPosition] = true
-		parentNodes = append(parentNodes, currentStep)
-		expandenNodes++
-
-		e.agent.position = currentStep
-		agentPerception := Percept(e.agent, e.board)
-
-		for _, perception := range agentPerception {
-			if !visited[perception.Coordinate] {
-				movementCost := float32(currentStep.Cost) + getCellCost(e.board[perception.Coordinate.X][perception.Coordinate.Y])
-				priorityQueue.Push(
-					datatypes.Element[datatypes.AgentStep]{
-						Value: datatypes.AgentStep{
-							Depth:            currentStep.Depth + 1,
-							Action:           perception.Direction,
-							Cost:             int(movementCost), // Use the calculated movement cost
-							CurrentPosition:  perception.Coordinate,
-							PreviousPosition: currentStep.CurrentPosition,
-						},
-						Priority: int(movementCost), // Priority is based on the total cost to reach this node
-					},
-				)
-			}
-		}
-	}
-
-	// If no path found, return empty result
-	return []datatypes.BoardCoordinate{}, expandenNodes, 0
 }
 
 func StartSearch(strategy int, scannedMatrix datatypes.ScannedMatrix) datatypes.SearchResult {
